refactor a bit, store wallets using deterministic bdk checksum name instead of literal descriptor or only external checksum
