+++ src/wallet.rs
@@ -1,6 +1,22 @@
-use bdk::{bitcoin::Txid, TransactionDetails};
+use bdk::{
+    bitcoin::{
+        secp256k1::{All, Secp256k1},
+        Network, Txid,
+    },
+    chain::{keychain::LocalChangeSet, ConfirmationTimeAnchor},
+    wallet::wallet_name_from_descriptor,
+    KeychainKind, TransactionDetails, Wallet,
+};
+use bdk_esplora::{esplora_client, EsploraAsyncExt};
+use bdk_file_store::Store;
+use cln_plugin::Error;
+use home::home_dir;
 use serde::{Deserialize, Serialize};
-use std::{collections::BTreeMap, fmt};
+use std::{collections::BTreeMap, fmt, io::Write};
+
+pub const DATADIR: &str = ".watchdescriptor";
+const STOP_GAP: usize = 50;
+const PARALLEL_REQUESTS: usize = 5;
 
 /// Errors related to the `watchdescriptor` command.
 #[derive(Debug)]
@@ -34,6 +50,7 @@ pub struct DescriptorWallet {
     // pub last_synced: Option<BlockTime>,
     // #[serde(skip_serializing, skip_deserializing)]
     pub transactions: BTreeMap<Txid, TransactionDetails>,
+    pub network: Option<Network>,
 }
 impl DescriptorWallet {
     fn new(
@@ -41,6 +58,7 @@ impl DescriptorWallet {
         change_descriptor: Option<&str>,
         birthday: Option<u64>,
         gap: Option<u64>,
+        network: Option<Network>,
     ) -> Result<Self, WatchError> {
         let mut params = DescriptorWallet::from_descriptor(descriptor)?;
         if change_descriptor.is_some() {
@@ -51,6 +69,9 @@ impl DescriptorWallet {
         }
         if gap.is_some() {
             params = params.with_gap(gap.unwrap())?
+        }
+        if network.is_some() {
+            params = params.with_network(network.unwrap())?
         }
         Ok(params)
     }
@@ -63,6 +84,7 @@ impl DescriptorWallet {
             gap: None,
             // last_synced: None,
             transactions: BTreeMap::new(),
+            network: None,
         })
     }
 
@@ -105,6 +127,13 @@ impl DescriptorWallet {
         }
     }
 
+    fn with_network(self, network: Network) -> Result<Self, WatchError> {
+        Ok(Self {
+            network: Some(network),
+            ..self
+        })
+    }
+
     // pub fn update_last_synced(&mut self, last_synced: BlockTime) {
     //     self.last_synced = Some(last_synced);
     // }
@@ -122,6 +151,104 @@ impl DescriptorWallet {
         }
         new_txs
         // self.transactions = transactions;
+    }
+
+    pub fn get_name(&self) -> Result<String, Error> {
+        Ok(wallet_name_from_descriptor(
+            &self.descriptor,
+            self.change_descriptor.as_ref(),
+            self.network.unwrap(),
+            &Secp256k1::<All>::new(),
+        )?)
+    }
+
+    pub async fn fetch_wallet<'a>(
+        &self,
+        // dw: &DescriptorWallet,
+    ) -> Result<Wallet<Store<'a, LocalChangeSet<KeychainKind, ConfirmationTimeAnchor>>>, Error>
+    {
+        // let db_path = std::env::temp_dir().join("bdk-esplora-async-example");
+        log::info!("creating path");
+        // let db_filename: String = general_purpose::STANDARD_NO_PAD.encode(dw.descriptor.as_bytes());
+        // let db_filename: String = calc_checksum(&dw.descriptor)?;
+        let db_filename = self.get_name()?;
+        let db_path = home_dir()
+            .unwrap()
+            .join(DATADIR)
+            .join(format!("{}.db", db_filename,));
+        log::info!("searching for path: {:?}", db_path);
+        let db = Store::<bdk::wallet::ChangeSet>::new_from_path(DATADIR.as_bytes(), db_path)?;
+        log::info!("db created!");
+        // let external_descriptor = "wpkh(tprv8ZgxMBicQKsPdy6LMhUtFHAgpocR8GC6QmwMSFpZs7h6Eziw3SpThFfczTDh5rW2krkqffa11UpX3XkeTTB2FvzZKWXqPY54Y6Rq4AQ5R8L/84'/0'/0'/0/*)";
+        // mutinynet_descriptor = "wpkh(tprv8ZgxMBicQKsPdSAgthqLZ5ZWQkm5As4V3qNA5G8KKxGuqdaVVtBhytrUqRGPm4RxTktSdvch8JyUdfWR8g3ddrC49WfZnj4iGZN8y5L8NPZ/*)"
+        let _mutinynet_descriptor_ext = "wpkh(tprv8ZgxMBicQKsPdSAgthqLZ5ZWQkm5As4V3qNA5G8KKxGuqdaVVtBhytrUqRGPm4RxTktSdvch8JyUdfWR8g3ddrC49WfZnj4iGZN8y5L8NPZ/84'/0'/0'/0/*)";
+        let _mutinynet_descriptor_int = "wpkh(tprv8ZgxMBicQKsPdSAgthqLZ5ZWQkm5As4V3qNA5G8KKxGuqdaVVtBhytrUqRGPm4RxTktSdvch8JyUdfWR8g3ddrC49WfZnj4iGZN8y5L8NPZ/84'/0'/0'/1/*)";
+        let _mutinynet_descriptor_ext_2 = "wpkh(tprv8ZgxMBicQKsPeRye8MhHA8hLxMuomycmGYXyRs7zViNck2VJsCJMTPt81Que8qp3PyPgQRnN7Gb1JyBVBKgj8AKEoEmmYxYDwzZJ63q1yjA/84'/0'/0'/0/*)";
+        let _mutinynet_descriptor_int_2 = "wpkh(tprv8ZgxMBicQKsPeRye8MhHA8hLxMuomycmGYXyRs7zViNck2VJsCJMTPt81Que8qp3PyPgQRnN7Gb1JyBVBKgj8AKEoEmmYxYDwzZJ63q1yjA/84'/0'/0'/1/*)";
+        // let external_descriptor = "wpkh(tpubEBr4i6yk5nf5DAaJpsi9N2pPYBeJ7fZ5Z9rmN4977iYLCGco1VyjB9tvvuvYtfZzjD5A8igzgw3HeWeeKFmanHYqksqZXYXGsw5zjnj7KM9/*)";
+        // let internal_descriptor = "wpkh(tprv8ZgxMBicQKsPdy6LMhUtFHAgpocR8GC6QmwMSFpZs7h6Eziw3SpThFfczTDh5rW2krkqffa11UpX3XkeTTB2FvzZKWXqPY54Y6Rq4AQ5R8L/84'/0'/0'/1/*)";
+
+        // let external_descriptor = mutinynet_descriptor_ext;
+        // let internal_descriptor = mutinynet_descriptor_int;
+        let external_descriptor = self.descriptor.clone();
+        let internal_descriptor = self.change_descriptor.clone();
+        // log::info!(
+        //     "about to create wallet {}, {:?}",
+        //     &dw.descriptor,
+        //     &dw.change_descriptor,
+        // );
+        let mut wallet = Wallet::new(
+            &external_descriptor,
+            internal_descriptor.as_ref(),
+            db,
+            Network::Testnet,
+        )?;
+        log::info!("wallet created!");
+
+        // let address = wallet.get_address(AddressIndex::New);
+        // log::info!("Generated Address: {}", address);
+
+        let balance = wallet.get_balance();
+        log::info!("Wallet balance before syncing: {} sats", balance.total());
+
+        log::info!("Syncing...");
+        let client =
+            // esplora_client::Builder::new("https://blockstream.info/testnet/api").build_async()?;
+            esplora_client::Builder::new("https://mutinynet.com/api").build_async()?;
+
+        let local_chain = wallet.checkpoints();
+        let keychain_spks = wallet
+            .spks_of_all_keychains()
+            .into_iter()
+            .map(|(k, k_spks)| {
+                let mut once = Some(());
+                let mut stdout = std::io::stdout();
+                let k_spks = k_spks
+                    .inspect(move |(spk_i, _)| match once.take() {
+                        Some(_) => log::info!("\nScanning keychain [{:?}]", k),
+                        None => log::info!(" {:<3}", spk_i),
+                    })
+                    .inspect(move |_| stdout.flush().expect("must flush"));
+                (k, k_spks)
+            })
+            .collect();
+        log::info!("CAG finished scanning");
+        let update = client
+            .scan(
+                local_chain,
+                keychain_spks,
+                [],
+                [],
+                STOP_GAP,
+                PARALLEL_REQUESTS,
+            )
+            .await?;
+        wallet.apply_update(update)?;
+        wallet.commit()?;
+
+        let balance = wallet.get_balance();
+        log::info!("Wallet balance after syncing: {} sats", balance.total());
+        return Ok(wallet);
     }
 }
 
@@ -157,7 +284,7 @@ impl TryFrom<serde_json::Value> for DescriptorWallet {
                             None
                         };
 
-                        DescriptorWallet::new(descriptor, change_descriptor, birthday, gap)
+                        DescriptorWallet::new(descriptor, change_descriptor, birthday, gap, None)
                     }
                     _ => Err(WatchError::InvalidFormat(format!("Unexpected request format. The request needs 1-4 parameters. Received: {param_count}"))),
                 }
@@ -188,6 +315,7 @@ impl TryFrom<serde_json::Value> for DescriptorWallet {
                             Some(v) => Some(v.as_u64().unwrap()),
                             None => None,
                         },
+                        None,
                     )
                 }
             },
